+++ src/formats/NinSnesInstr.h
@@ -39,7 +39,7 @@ public:
 	virtual bool LoadInstr();
 
 	static bool IsValidHeader(RawFile * file, NinSnesVersion version, uint32_t addrInstrHeader, uint32_t spcDirAddr);
-	static size_t ExpectedSize(NinSnesVersion version);
+	static uint32_t ExpectedSize(NinSnesVersion version);
 
 	NinSnesVersion version;
